Remove the compile-time bus ceiling entirely
max_buses was runtime but still capped by AMY_MAX_BUSES. Two things
forced that cap; both are now gone, so max_buses is bounded only by RAM.

The real blocker was the params enum. Volume was the one bus-directed
parameter that encoded its bus in the param id -- VOLUME_BASE + bus, one
id per bus -- so the bus count was limited by how many ids fit before
MODE (99). Every other bus-directed param (EQ_*, ECHO_*, CHORUS_*,
REVERB_*) already names its bus in delta.osc, which is a uint16_t. Volume
now does the same: a single VOLUME id, one delta per bus, bus in osc.
That drops the 33-bus wall and frees ids 66..98. "V1,0.5" still sets
consecutive buses; each entry is just its own delta now.

The rest was fixed-width tables, which now allocate from max_buses:
global_state.bus/volume, fbl and per_osc_fb (per-core arrays of bus
pointers), and the per-block volume_scale, which was a stack array sized
by the ceiling and is now allocated once. bus_gain went with them -- it
was declared and never read.

Bus numbers were uint8_t in synthinfo, mod_synthinfo, amy_event,
instrument_info, config.max_buses, highest_bus and every config_*/filter
signature, which would have silently truncated bus 300 to 44. All widened
to uint16_t, including the public amy_external_bus_postprocess_hook.

amy_event keeps a fixed volume vector, now AMY_MAX_VOLUME_LIST (8) rather
than a bus count: it bounds how many buses one V message can set, the way
MAX_ALGO_OPS bounds a message's operator list. It is not a limit on buses
-- any bus is reachable with bus=n, volume=x -- and the event has to stay
a value type, so this one stays compile-time.

tests/test_bus_config.c now runs 300 buses, checks the volume vector
lands on the right buses from a non-zero start and stops at the last one,
and renders 64 buses with FX. The 124 WAV tests are bit-identical,
including the two that use volume='2,0.5'.

+++ src/examples.c
@@ -177,14 +177,14 @@ void example_patches() {
 }
 void example_reverb() {
     if(AMY_HAS_REVERB) {
-        uint8_t bus = 0;
+        uint16_t bus = 0;
         config_reverb(bus, 2, REVERB_DEFAULT_LIVENESS, REVERB_DEFAULT_DAMPING, REVERB_DEFAULT_XOVER_HZ); 
     }
 }
 
 void example_chorus() {
     if(AMY_HAS_CHORUS) {
-        uint8_t bus = 0;
+        uint16_t bus = 0;
         config_chorus(bus, 0.8, CHORUS_DEFAULT_MAX_DELAY, CHORUS_DEFAULT_LFO_FREQ, CHORUS_DEFAULT_MOD_DEPTH);
     }
 }
